+++ src/DisplayGame.js
@@ -3,22 +3,22 @@ function DisplayGame(props) {
     return(
         <section>
 
-            {!props.game ? (
+            {!props.games ? (
                 <h2>Whoops, something went wrong. Try refreshing.</h2>
             ) :(
                 <>
                 <h2>Check out this game!</h2>
                 <div className="gameCard">
-                    {props.map( (game) => {
-                        return(
-                            
-                            <div className="gameContainer" key={game.id}>
-                                <h2>{game.title}</h2>
-                                <img src={game.thumbnail} alt={game.title} />
-                                <p>{game.short_description}</p>
+                    {/* ternarary here */}
+                    {/* {props.games( (game) => { */}
+                        
+                            <div className="gameContainer" key={props.games.id}>
+                                <h2>{props.games.title}</h2>
+                                <img src={props.games.thumbnail} alt={props.games.title} />
+                                <p>{props.games.short_description}</p>
                             </div>
-                        )
-                    })}
+                        
+                    {/* })} */}
                 </div>
                 </>
             )}
